Break out GFA writer functions
